+++ kernel/syscall/path.c
@@ -28,8 +28,8 @@ struct path* path_from_dirfd(int dirfd) {
     return path_dup(file->path);
 }
 
-NODISCARD static int open(const struct path* base, const char* user_pathname,
-                          int flags, unsigned mode) {
+NODISCARD static long open(const struct path* base, const char* user_pathname,
+                           int flags, unsigned mode) {
     char pathname[PATH_MAX];
     int rc = copy_pathname_from_user(pathname, user_pathname);
     if (IS_ERR(rc))
@@ -45,24 +45,24 @@ NODISCARD static int open(const struct path* base, const char* user_pathname,
     return files_alloc_fd(current->files, -1, file);
 }
 
-int sys_open(const char* user_pathname, int flags, unsigned mode) {
+long sys_open(const char* user_pathname, int flags, unsigned mode) {
     SCOPED_LOCK(fs, current->fs);
     return open(current->fs->cwd, user_pathname, flags, mode);
 }
 
-int sys_openat(int dirfd, const char* user_pathname, int flags, mode_t mode) {
+long sys_openat(int dirfd, const char* user_pathname, int flags, mode_t mode) {
     struct path* base FREE(path) = path_from_dirfd(dirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
     return open(base, user_pathname, flags, mode);
 }
 
-int sys_creat(const char* user_pathname, mode_t mode) {
+long sys_creat(const char* user_pathname, mode_t mode) {
     return sys_open(user_pathname, O_CREAT | O_WRONLY | O_TRUNC, mode);
 }
 
-NODISCARD static int mknod(const struct path* base, const char* user_pathname,
-                           mode_t mode, dev_t dev) {
+NODISCARD static long mknod(const struct path* base, const char* user_pathname,
+                            mode_t mode, dev_t dev) {
     switch (mode & S_IFMT) {
     case S_IFREG:
     case S_IFCHR:
@@ -86,20 +86,20 @@ NODISCARD static int mknod(const struct path* base, const char* user_pathname,
     return 0;
 }
 
-int sys_mknod(const char* user_pathname, mode_t mode, dev_t dev) {
+long sys_mknod(const char* user_pathname, mode_t mode, dev_t dev) {
     SCOPED_LOCK(fs, current->fs);
     return mknod(current->fs->cwd, user_pathname, mode, dev);
 }
 
-int sys_mknodat(int dirfd, const char* user_pathname, mode_t mode, dev_t dev) {
+long sys_mknodat(int dirfd, const char* user_pathname, mode_t mode, dev_t dev) {
     struct path* base FREE(path) = path_from_dirfd(dirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
     return mknod(base, user_pathname, mode, dev);
 }
 
-NODISCARD static int mkdir(const struct path* base, const char* user_pathname,
-                           mode_t mode) {
+NODISCARD static long mkdir(const struct path* base, const char* user_pathname,
+                            mode_t mode) {
     char pathname[PATH_MAX];
     int rc = copy_pathname_from_user(pathname, user_pathname);
     if (IS_ERR(rc))
@@ -111,20 +111,20 @@ NODISCARD static int mkdir(const struct path* base, const char* user_pathname,
     return 0;
 }
 
-int sys_mkdir(const char* user_pathname, mode_t mode) {
+long sys_mkdir(const char* user_pathname, mode_t mode) {
     SCOPED_LOCK(fs, current->fs);
     return mkdir(current->fs->cwd, user_pathname, mode);
 }
 
-int sys_mkdirat(int dirfd, const char* user_pathname, mode_t mode) {
+long sys_mkdirat(int dirfd, const char* user_pathname, mode_t mode) {
     struct path* base FREE(path) = path_from_dirfd(dirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
     return mkdir(base, user_pathname, mode);
 }
 
-NODISCARD static int access(const struct path* base, const char* user_pathname,
-                            int mode) {
+NODISCARD static long access(const struct path* base, const char* user_pathname,
+                             int mode) {
     (void)mode; // File permissions are not implemented in this system.
 
     char pathname[PATH_MAX];
@@ -137,20 +137,20 @@ NODISCARD static int access(const struct path* base, const char* user_pathname,
     return 0;
 }
 
-int sys_access(const char* user_pathname, int mode) {
+long sys_access(const char* user_pathname, int mode) {
     SCOPED_LOCK(fs, current->fs);
     return access(current->fs->cwd, user_pathname, mode);
 }
 
-int sys_faccessat(int dirfd, const char* user_pathname, int mode) {
+long sys_faccessat(int dirfd, const char* user_pathname, int mode) {
     struct path* base FREE(path) = path_from_dirfd(dirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
     return access(base, user_pathname, mode);
 }
 
-NODISCARD static int link(struct inode* old_inode, const struct path* new_base,
-                          const char* user_newpath) {
+NODISCARD static long link(struct inode* old_inode, const struct path* new_base,
+                           const char* user_newpath) {
     if (S_ISDIR(old_inode->mode))
         return -EPERM;
 
@@ -171,7 +171,7 @@ NODISCARD static int link(struct inode* old_inode, const struct path* new_base,
     return inode_link(new_path->parent->inode, new_path->basename, old_inode);
 }
 
-int sys_link(const char* user_oldpath, const char* user_newpath) {
+long sys_link(const char* user_oldpath, const char* user_newpath) {
     SCOPED_LOCK(fs, current->fs);
     struct path* old_path FREE(path) = vfs_resolve_path_at(
         current->fs->cwd, user_oldpath, O_NOFOLLOW | O_NOFOLLOW_NOERROR);
@@ -180,8 +180,8 @@ int sys_link(const char* user_oldpath, const char* user_newpath) {
     return link(old_path->inode, current->fs->cwd, user_newpath);
 }
 
-int sys_linkat(int olddirfd, const char* user_oldpath, int newdirfd,
-               const char* user_newpath, int flags) {
+long sys_linkat(int olddirfd, const char* user_oldpath, int newdirfd,
+                const char* user_newpath, int flags) {
     if (flags & ~(AT_SYMLINK_FOLLOW | AT_EMPTY_PATH))
         return -EINVAL;
 
@@ -224,7 +224,7 @@ int sys_linkat(int olddirfd, const char* user_oldpath, int newdirfd,
 }
 
 NODISCARD
-static int unlink(const struct path* base, const char* user_pathname) {
+static long unlink(const struct path* base, const char* user_pathname) {
     char pathname[PATH_MAX];
     int rc = copy_pathname_from_user(pathname, user_pathname);
     if (IS_ERR(rc))
@@ -239,12 +239,13 @@ static int unlink(const struct path* base, const char* user_pathname) {
     return inode_unlink(path->parent->inode, path->basename);
 }
 
-int sys_unlink(const char* user_pathname) {
+long sys_unlink(const char* user_pathname) {
     SCOPED_LOCK(fs, current->fs);
     return unlink(current->fs->cwd, user_pathname);
 }
 
-NODISCARD static int rmdir(const struct path* base, const char* user_pathname) {
+NODISCARD
+static long rmdir(const struct path* base, const char* user_pathname) {
     char pathname[PATH_MAX];
     int rc = copy_pathname_from_user(pathname, user_pathname);
     if (IS_ERR(rc))
@@ -263,7 +264,7 @@ NODISCARD static int rmdir(const struct path* base, const char* user_pathname) {
     return inode_unlink(path->parent->inode, path->basename);
 }
 
-int sys_unlinkat(int dirfd, const char* user_pathname, int flags) {
+long sys_unlinkat(int dirfd, const char* user_pathname, int flags) {
     if (flags & ~AT_REMOVEDIR)
         return -EINVAL;
 
@@ -274,14 +275,14 @@ int sys_unlinkat(int dirfd, const char* user_pathname, int flags) {
                                   : unlink(base, user_pathname);
 }
 
-int sys_rmdir(const char* user_pathname) {
+long sys_rmdir(const char* user_pathname) {
     SCOPED_LOCK(fs, current->fs);
     return rmdir(current->fs->cwd, user_pathname);
 }
 
 NODISCARD
-static int rename(const struct path* old_base, const char* user_oldpath,
-                  const struct path* new_base, const char* user_newpath) {
+static long rename(const struct path* old_base, const char* user_oldpath,
+                   const struct path* new_base, const char* user_newpath) {
     char old_pathname[PATH_MAX];
     int rc = copy_pathname_from_user(old_pathname, user_oldpath);
     if (IS_ERR(rc))
@@ -330,19 +331,19 @@ static int rename(const struct path* old_base, const char* user_oldpath,
     return inode_unlink(old_path->parent->inode, old_path->basename);
 }
 
-int sys_rename(const char* user_oldpath, const char* user_newpath) {
+long sys_rename(const char* user_oldpath, const char* user_newpath) {
     SCOPED_LOCK(fs, current->fs);
     return rename(current->fs->cwd, user_oldpath, current->fs->cwd,
                   user_newpath);
 }
 
-int sys_renameat(int olddirfd, const char* user_oldpath, int newdirfd,
-                 const char* user_newpath) {
+long sys_renameat(int olddirfd, const char* user_oldpath, int newdirfd,
+                  const char* user_newpath) {
     return sys_renameat2(olddirfd, user_oldpath, newdirfd, user_newpath, 0);
 }
 
-int sys_renameat2(int olddirfd, const char* user_oldpath, int newdirfd,
-                  const char* user_newpath, unsigned int flags) {
+long sys_renameat2(int olddirfd, const char* user_oldpath, int newdirfd,
+                   const char* user_newpath, unsigned int flags) {
     if (flags)
         return -EINVAL;
 
@@ -357,8 +358,8 @@ int sys_renameat2(int olddirfd, const char* user_oldpath, int newdirfd,
     return rename(old_base, user_oldpath, new_base, user_newpath);
 }
 
-NODISCARD static int symlink(const struct path* base, const char* user_target,
-                             const char* user_linkpath) {
+NODISCARD static long symlink(const struct path* base, const char* user_target,
+                              const char* user_linkpath) {
     char target[PATH_MAX];
     int rc = copy_pathname_from_user(target, user_target);
     if (IS_ERR(rc))
@@ -380,13 +381,13 @@ NODISCARD static int symlink(const struct path* base, const char* user_target,
     return file_symlink(file, target);
 }
 
-int sys_symlink(const char* user_target, const char* user_linkpath) {
+long sys_symlink(const char* user_target, const char* user_linkpath) {
     SCOPED_LOCK(fs, current->fs);
     return symlink(current->fs->cwd, user_target, user_linkpath);
 }
 
-int sys_symlinkat(const char* user_target, int newdirfd,
-                  const char* user_linkpath) {
+long sys_symlinkat(const char* user_target, int newdirfd,
+                   const char* user_linkpath) {
     struct path* base FREE(path) = path_from_dirfd(newdirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
@@ -394,8 +395,8 @@ int sys_symlinkat(const char* user_target, int newdirfd,
 }
 
 NODISCARD
-static int readlink(const struct path* base, const char* user_pathname,
-                    char* user_buf, size_t bufsiz) {
+static long readlink(const struct path* base, const char* user_pathname,
+                     char* user_buf, size_t bufsiz) {
     char pathname[PATH_MAX];
     int rc = copy_pathname_from_user(pathname, user_pathname);
     if (IS_ERR(rc))
@@ -421,13 +422,13 @@ static int readlink(const struct path* base, const char* user_pathname,
     return len;
 }
 
-ssize_t sys_readlink(const char* user_pathname, char* user_buf, size_t bufsiz) {
+long sys_readlink(const char* user_pathname, char* user_buf, size_t bufsiz) {
     SCOPED_LOCK(fs, current->fs);
     return readlink(current->fs->cwd, user_pathname, user_buf, bufsiz);
 }
 
-ssize_t sys_readlinkat(int dirfd, const char* user_pathname, char* user_buf,
-                       size_t bufsiz) {
+long sys_readlinkat(int dirfd, const char* user_pathname, char* user_buf,
+                    size_t bufsiz) {
     struct path* base FREE(path) = path_from_dirfd(dirfd);
     if (IS_ERR(ASSERT(base)))
         return PTR_ERR(base);
